ViTamin/requirements-eval.txt
absl-py==1.3.0
aiohttp==3.8.3
aioitertools==0.11.0
aiosignal==1.3.1
albumentations==1.3.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
azure-core==1.26.1
azure-storage-blob==12.14.1
beautifulsoup4==4.11.1
braceexpand==0.1.7
cachetools==5.2.0
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
clip-benchmark==1.4.0
cloudpathlib==0.13.0
colorama==0.4.4
contourpy==1.0.7
cryptography==39.0.2
cycler==0.11.0
dataclasses==0.6
datasets==2.8.0
dill==0.3.6
dm-tree==0.1.7
docker-pycreds==0.4.0
docutils==0.16
etils==0.9.0
exifread-nocycle==3.0.1
# pybind11==2.10.1
# fasttext==0.9.2
# fasttext-langdetect==1.0.3
filelock==3.8.0
fire==0.4.0
flatbuffers==22.10.26
fonttools==4.39.2
frozenlist==1.3.3
fsspec==2022.11.0
ftfy==6.1.1
gast==0.4.0
# gcld3==3.0.13
gdown==4.5.3
gitdb==4.0.9
gitpython==3.1.29
google-api-core==2.11.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.50.0
h5py==3.7.0
huggingface-hub==0.11.0
idna==3.4
imageio==2.22.4
img2dataset==1.40.0
importlib-resources==5.10.0
isodate==0.6.1
jmespath==1.0.1
joblib==1.2.0
kaleido==0.2.1
keras==2.11.0
keras-preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
littleutils==0.2.2
markdown==3.4.1
markupsafe==2.1.1
matplotlib==3.7.1
mock==4.0.3
msrest==0.7.1
multidict==6.0.3
multiprocess==0.70.14
networkx==2.8.8
nltk==3.8.1
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
ogb==1.3.5
# open-clip-torch==2.16.1
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
outdated==0.2.2
packaging==21.3
pandas==1.5.1
pathtools==0.1.2
patsy==0.5.3
pillow==9.3.0
plotly==5.13.1
promise==2.3
protobuf==3.20.2
psutil==5.9.4
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycld3==0.22
pycparser==2.21
pyparsing==3.0.9
pysimdjson==5.0.2
pysocks==1.7.1
python-dateutil==2.8.2
pytz==2022.6
pywavelets==1.4.1
pyyaml==5.4.1
qudida==0.0.4
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.7.2
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
sentencepiece==0.1.97
sentry-sdk==1.11.1
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
statsmodels==0.13.5
tenacity==8.2.2
termcolor==2.1.0
threadpoolctl==3.1.0
tifffile==2022.10.10
# timm==0.6.11
tokenizers==0.13.2
toml==0.10.2
# torch==1.13.0
# torchvision==0.14.0
tqdm==4.64.1
transformers==4.24.0
typeguard==2.13.3
typing-extensions==4.4.0
urllib3==1.26.12
wandb==0.12.21
wcwidth==0.2.5
webdataset==0.2.31
werkzeug==2.2.2
wget==3.2
wilds==2.0.0
wrapt==1.14.1
xxhash==3.2.0
yarl==1.8.2
zipp==3.10.0
